* Need to expose tool locations, as we need some specific variants for GNU wc, etc * fix some use of nounset * Allow bash >= 4 * expose install.sh --version and --baseurl * allow install.sh from local directory

+++ install.sh
@@ -58,27 +58,38 @@ function __main {
   # resolve this script name
   basename=$(basename $0)
 
+  # when fetched remotely avoid confusion with detcted basename
+  if [ "$basename" = "bash" ]; then
+    basename='install.sh'
+  fi
+
   # determine fully-qualified base directory
   basedir=$(dirname $0)
   basedir=$(cd "$basedir" && pwd)
+
+  # stable defaults for usage display
+  local default_verbose=${verbose}
+  local default_baseurl='https://github.com/jdillon/commando-bash'
+  local default_version='master'
+
+  # configurable options
+  local baseurl=${default_baseurl}
+  local version=${default_version}
 
   # display usage and exit
   function usage {
     printf "\nusage: $basename [options]
 
 options:
-  -h,--help       Show usage
-  -v,--verbose    Verbose output
-  --              Stop processing options
+  -h,--help             Show usage
+  -v,--verbose          Verbose output; default: ${default_verbose}
+  --version <version>   Select version; default: ${default_version}
+  --baseurl <url>       Select base URL; default: ${default_baseurl}
+  --                    Stop processing options
 \n"
 
     exit 2
   }
-
-  cd "$basedir"
-
-  local baseurl='https://github.com/jdillon/commando-bash'
-  local version='master'
 
   # parse options and collect arguments
   local -a arguments
@@ -90,8 +101,16 @@ options:
         usage
         ;;
       -v|--verbose)
-        verbose=true
         shift
+        verbose=true
+        ;;
+      --baseurl)
+        baseurl="$2"
+        shift 2
+        ;;
+      --version)
+        version="$2"
+        shift 2
         ;;
       -*)
         die "Unknown option: $opt"
@@ -116,6 +135,7 @@ options:
 
   local disturl="$baseurl/archive/${version}.tar.gz"
   local tmpdir=`mktemp -d`
+  local projectdir=$(pwd)
 
   if ${verbose}; then
     log "Bash: $BASH $BASH_VERSINFO $BASH_VERSION"
@@ -125,29 +145,45 @@ options:
     log "Version: $version"
     log "Distribution URL: $disturl"
     log "Temp directory: $tmpdir"
+    log "Project directory: $projectdir"
+    log "Arguments: ${arguments[@]}"
   fi
 
-  local distfile="$tmpdir/dist.tgz"
-  log "Distribution archive: $distfile"
-  curl --location --silent --output "$distfile" "$disturl"
+  if [[ ${baseurl} == file:* ]]; then
+    # use local release distribution
+    local releasedir=${baseurl:5}
+    if [ ! -d ${releasedir} ]; then
+      die "Invalid baseurl; not a directory: ${releasedir}"
+    fi
+  else
+    # fetch release distribution
+    local distfile="$tmpdir/dist.tgz"
+    local distdir="$tmpdir/dist"
+    local releasedir="$distdir/commando-bash-$version"
 
-  local distdir="$tmpdir/dist"
-  log "Distribution directory: $distdir"
-  mkdir "$distdir"
-  tar -xzf "$distfile" -C "$distdir"
+    log "Distribution archive: $distfile"
+    curl --location --silent --output "$distfile" "$disturl"
 
-  local releasedir="$distdir/commando-bash-$version"
+    log "Distribution directory: $distdir"
+    mkdir "$distdir"
+    tar -xzf "$distfile" -C "$distdir"
+  fi
+
   log "Release directory: $releasedir"
+
+  local setup="$releasedir/setup.sh"
+  if [ ! -f ${setup} ]; then
+    die "Invalid release directory; missing: ${setup}"
+  fi
 
   if ${verbose}; then
     log '----8<----'
-    find "$releasedir"
+    find "$releasedir" -type f ! -path "*/.git*"
     log '---->8----'
   fi
 
-  local setup="$releasedir/setup.sh"
   source "$setup"
-  __setup "$releasedir" "$basedir"
+  __setup "$releasedir" "$projectdir"
 
   rm -rf "$tmpdir"
 }
@@ -159,7 +195,7 @@ options:
 __output_helpers
 
 # compatibility check
-if [ "$BASH_VERSINFO" != '4' ]; then
+if [ "$BASH_VERSINFO" -lt '4' ]; then
   die "Incompatible Bash detected: $BASH $BASH_VERSINFO $BASH_VERSION"
 fi
